some compile fix
git-svn-id: file:///svn/rdup/trunk@1753 0f4cb41f-7f08-0410-98f9-8ec83d93c06f

+++ gfunc.c
@@ -179,7 +179,7 @@ entry_print_data(FILE *out, char n, struct r_entry *e)
 			/* only the name in case of soft- or hardlinks
 			 * filesize has the length what we should print */
 			if (S_ISLNK(e->f_mode) || e->f_lnk == 1) {
-				if (fwrite(e->f_name, e->f_size, sizeof(char), out) != e->f_size) {
+				if (fwrite(e->f_name, e->f_size, sizeof(char), out) != (size_t)e->f_size) {
 					msg(_("Write failure: %s"), strerror(errno));
 				
 				}
